Fix `make_distributed_plan` aggregation cutoff settings
Apply nested query `SETTINGS` changes before deriving the aggregate cutoff, and fix the Darwin `size_t`/`UInt64` mismatch in `AggregatingTransform`.

https://github.com/ClickHouse/ClickHouse/pull/114690

https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=114690&sha=c7bf5e0f420b3c6229b6aabdbb6b9c697f8b9bc2&name_0=PR&name_1=Fast%20test

+++ src/Processors/Transforms/AggregatingTransform.cpp
@@ -1071,7 +1071,7 @@ Block buildKeptKeysSeedBlock(
         if (!rows_left)
             break;
         chassert(!agg_chunk.is_overflows);
-        size_t rows_to_take = std::min(rows_left, agg_chunk.chunk.getNumRows());
+        size_t rows_to_take = std::min(rows_left, static_cast<size_t>(agg_chunk.chunk.getNumRows()));
         const auto & columns = agg_chunk.chunk.getColumns();
         for (size_t i = 0; i < keys_size; ++i)
             seed_columns[i]->insertRangeFrom(*columns[i], 0, rows_to_take);
